Make Arm postions with a nested arm position class
