+++ src/main/java/frc/robot/subsystems/coralManipulator/states/PositionState.java
@@ -5,7 +5,6 @@
 package frc.robot.subsystems.coralManipulator.states;
 
 import frc.robot.subsystems.coralManipulator.CoralManipulator;
-import frc.robot.subsystems.coralManipulator.states.*;
 import frc.robot.utils.commandUtils.State;
 
 public class PositionState extends State<CoralManipulator> {
